requirements.txt
bcbio-gff
biopython
requests
junit-xml
apollo>=2.0
xunit-wrapper
